PDP: Refactor request to share initialization and error checking.
Each session request starts with similar logic to check for the
existence of a user and that the user has the requested roles. This
changeset adds a PDP.request() function that handles this checking
and returns the generated data structures for the user's Roles and
UserGraph, the induced subgraph of the user.

Note: Cpol is unable to use the shared logic because CPOL uses a
different method to validate users.