HWDS: Add HWDS-based SDP.
The HWDS-based SDP is similar to the MapBitSet. The PDP is the same
for now between the two. The HWDS SDP uses a HWDS to store the map
between sessions and permission bit sets. The permission bit sets
are stored in the HWDS by their hashCode, and are stored in a secondary
HashMap data structure that maps the hashCode to the bit set. The
BitSet structure is wrapped within a new CountedBitSet class that
is used to track how many BitSet instances have been inserted - removed
in the HashMap.

Changes are also made to enable the new HWDS SDP in the Simulation driver
and wrapping support scripts for running the benchmark code.