experiments: refactor Simulate3 and Measurements
Make the within-VM invocation check itself for being finished. This
eliminates the step of calling Measurements after Simulate3 in the
experiment script. It also can permit the experiment to finish sooner,
since Simulate3 can terminate when the steady state is reached. Also
this change extends the number of iterations within the VM to a max of 30.